--- config/content/Hero.js
@@ -1,9 +1,0 @@
-/* eslint-disable max-len */
-export default {
-  id: 'home',
-  titleLine: { text: '' },
-  tagline: {
-    
-    taglinedesc: "Hop on the bandwagon, and join one of India’s largest Student-run hackathons. HackNITR is back with version 5.0, which is now fully offline!",
-  },
-};
