+++ config/index.js
@@ -1,3 +1,3 @@
 export { default as footer } from './content/Footer';
 export { default as website } from './website';
-export { default as hero } from './content/Hero';
+// export { default as hero } from './content/Hero';
