--- src/components/shared/DiscordBtn.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import { hero } from '../../../config';
-import "./DevfolioBtn.css"
-
-function DiscordButton() {
-  return (
-    
-    <div className="btn" ontouchstart="">
-        <div className="register_btn">
-          <button><a href="">D</a></button>
-        </div>
-      </div>
-    
-  );
-}
-
-export default DiscordButton;
